main coerce.js now exists
